requirements.txt
requests==2.31.0
python-dotenv==1.0.0
pydantic==2.5.2
asyncio==3.4.3
PyYAML==6.0.1